requirements.txt
anthropic>=0.40.0
PyYAML>=6.0.2
requests>=2.32.3
pytest>=8.3.4
pytest-mock>=3.14.0
ruff>=0.8.0
